feat: BlockMaterialWithDynModel
