+++ src/main/kotlin/com/github/trc/clayium/common/blocks/ItemBlockMaterial.kt
@@ -1,6 +1,7 @@
 package com.github.trc.clayium.common.blocks
 
 import com.github.trc.clayium.api.unification.ore.OrePrefix
+import com.github.trc.clayium.common.blocks.material.BlockMaterialBase
 import net.minecraft.item.ItemBlock
 import net.minecraft.item.ItemStack
 import net.minecraftforge.common.IRarity
