+++ src/main/kotlin/com/github/trc/clayium/common/blocks/material/BlockCompressedClay.kt
@@ -1,7 +1,6 @@
 package com.github.trc.clayium.common.blocks.material
 
 import com.github.trc.clayium.api.unification.material.CMaterial
-import com.github.trc.clayium.common.blocks.BlockMaterialBase
 import com.github.trc.clayium.common.blocks.properties.CMaterialProperty
 import com.github.trc.clayium.common.creativetab.ClayiumCTabs
 import net.minecraft.block.SoundType
