+++ src/main/kotlin/com/github/trc/clayium/common/blocks/material/BlockEnergizedClay.kt
@@ -2,7 +2,6 @@ package com.github.trc.clayium.common.blocks.material
 
 import com.github.trc.clayium.api.capability.ClayiumCapabilities
 import com.github.trc.clayium.api.unification.material.CMaterial
-import com.github.trc.clayium.common.blocks.BlockMaterialBase
 import com.github.trc.clayium.common.blocks.properties.CMaterialProperty
 import com.github.trc.clayium.common.creativetab.ClayiumCTabs
 import net.minecraft.block.SoundType
